fix start, added platofrms.go, cleanup
