+++ mesheryctl/internal/cli/root/system/config.go
@@ -40,16 +40,9 @@ import (
 // TODO: https://github.com/layer5io/me shery/issues/1022
 
 const paramName = "k8sfile"
-const contextName = "contextName"
 const kubeConfigYaml = "kubeconfig.yaml"
 
 var tokenPath string
-
-type k8sContext struct {
-	contextName    string
-	clusterName    string
-	currentContext bool
-}
 
 func getContexts(configFile, tokenPath string) ([]string, error) {
 	client := &http.Client{}
